requirements.txt
streamlit
pandas
sqlalchemy
python-dateutil